Bank-Service/pyproject.toml
[project]
name = "bank-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.3",
    "django>=5.2.1",
    "django-cors-headers>=4.7.0",
    "django-countries>=7.6.1",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "uvicorn[standard]>=0.34.2",
]
